refactor: 카드 정보 입력을 위한 CardInfoForm 컴포넌트로 분리
